Make buying/selling actually work

+++ src/stockmarketedu/SellStockServlet.java
@@ -29,7 +29,7 @@ public class SellStockServlet extends HttpServlet {
 			resp.sendRedirect("/student.jsp");
 			return;
 		}
-		int shares = Integer.parseInt(numShares);
+		double shares = Double.parseDouble(numShares);
 		
 		List<Supervisor> teachers = ObjectifyService.ofy().load().type(Supervisor.class).list();
 		// TODO: check for allowed stocks here
